+++ Sx4/src/main/java/com/sx4/bot/entities/economy/item/Pickaxe.java
@@ -1,8 +1,8 @@
 package com.sx4.bot.entities.economy.item;
 
-import java.util.List;
-
 import org.bson.Document;
+
+import java.util.List;
 
 public class Pickaxe extends Tool {
 
@@ -25,7 +25,7 @@ public class Pickaxe extends Tool {
 		);
 	}
 
-	public Pickaxe(String name, long price, long currentDurability, long maxDurability, List<ItemStack<Material>> craft, Material repairItem, long minYield, long maxYield, double multiplier) {
+	public Pickaxe(String name, long price, int currentDurability, int maxDurability, List<ItemStack<Material>> craft, Material repairItem, long minYield, long maxYield, double multiplier) {
 		super(name, price, ItemType.PICKAXE, currentDurability, maxDurability, craft, repairItem);
 		
 		this.minYield = minYield;
@@ -33,7 +33,7 @@ public class Pickaxe extends Tool {
 		this.multiplier = multiplier;
 	}
 	
-	public Pickaxe(String name, long price, long maxDurability, List<ItemStack<Material>> craft, Material repairItem, long minYield, long maxYield, double multiplier) {
+	public Pickaxe(String name, long price, int maxDurability, List<ItemStack<Material>> craft, Material repairItem, long minYield, long maxYield, double multiplier) {
 		this(name, price, maxDurability, maxDurability, craft, repairItem, minYield, maxYield, multiplier);
 	}
