+++ Sx4/src/main/java/com/sx4/bot/entities/economy/item/Tool.java
@@ -4,17 +4,17 @@ import java.util.List;
 
 public class Tool extends Item {
 	
-	private final long currentDurability;
-	private final long maxDurability;
+	private final int currentDurability;
+	private final int maxDurability;
 	
 	private final List<ItemStack<Material>> craft;
 	private final Material repairItem;
 
-	public Tool(String name, long price, ItemType type, long maxDurability, List<ItemStack<Material>> craft, Material repairItem) {
+	public Tool(String name, long price, ItemType type, int maxDurability, List<ItemStack<Material>> craft, Material repairItem) {
 		this(name, price, type, maxDurability, maxDurability, craft, repairItem);
 	}
 	
-	public Tool(String name, long price, ItemType type, long currentDurability, long maxDurability, List<ItemStack<Material>> craft, Material repairItem) {
+	public Tool(String name, long price, ItemType type, int currentDurability, int maxDurability, List<ItemStack<Material>> craft, Material repairItem) {
 		super(name, price, type);
 		
 		this.currentDurability = currentDurability;
@@ -23,11 +23,11 @@ public class Tool extends Item {
 		this.repairItem = repairItem;
 	}
 	
-	public long getCurrentDurability() {
+	public int getCurrentDurability() {
 		return this.currentDurability;
 	}
 	
-	public long getMaxDurability() {
+	public int getMaxDurability() {
 		return this.maxDurability;
 	}
